requirements.txt
asgiref==3.7.2
astroid==3.0.2
chardet==5.2.0
dill==0.3.7
Django==4.2.9
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.5
django-simple-history==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-standardized-errors==0.12.6
exceptiongroup==1.2.0
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.3.0
psycopg==3.1.16
psycopg-binary==3.1.16
PyJWT==2.8.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
reportlab==4.0.8
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
